Refactor details style.js
