+++ components/postPage/details/generalDetails/index.js
@@ -1,7 +1,7 @@
 import React from "react";
 import { Row, Col } from "antd";
 
-import { BoxOutline, Title, ItemKey, ItemValue } from "./style";
+import { BoxOutline, Title, ItemKey, ItemValue } from "../style";
 
 const index = ({ post }) => {
   return (
