+++ components/postPage/details/style.js
